requirements.txt
absl-py==1.4.0
anyio==3.7.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==7.1.2
colorama==0.4.6
Deprecated==1.2.14
exceptiongroup==1.1.3
ffmpeg-python==0.2.0
filelock==3.12.2
flatbuffers==1.12
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.57.0
h11==0.12.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.4
jaconv==0.3.4
Jinja2==3.1.2
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
llvmlite==0.40.1
Markdown==3.4.4
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
norbert==0.2.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
openai-whisper==20230314
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pykakasi==2.2.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
pytz==2023.3
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.11.2
six==1.16.0
sniffio==1.3.0
spleeter==2.4.0
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tiktoken==0.3.1
torch==2.0.1
tqdm==4.66.1
typer==0.3.2
typing_extensions==4.7.1
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
